fix: Set langpack option to addons-linter.createInstance
Since if langpack is 'undefined' in the config object for createInstance,
the ctor of ManifestJSONParser() calls getConfig() to get the default argument
value.  Thus getConfig() parses command line arguments (i.e. not the config
object which is passed to createInstance), then if the command which uses
cmd.lint() has no arguments, the command fails.