test: refactor failing test and improve coverage
A previous commit changed `src/cmd/index.js` to not directly export the
implementations, but to lazy-load those modules.
So the test that did a simple comparison started failing.

To fix this test and maintain test coverage, change the test to
temporarily intercept invocations of the original cmd module,
for every command that is exported in `src/cmd/index.js`.