fix: Stop bundling dependencies with slashes
External dependencies should not be bundled, because they are already
referenced via "dependencies" in package.json.

Before this patch, external modules with slashes in their identifier
were bundled anyway because there is no directory with that name in the
`node_modules` directory (which in turn prevents them from being added
to the `externals` object in the webpack config).

This is fixed by looking up dependencies in `package.json` instead.