fix: support multiple values after --ignore-files again
The yargs upgrade from 6.6.0 to 13.2.1 in edebef46eb broke the
documented ability to pass multiple files to ignore after the
`--ignore-files` (aka `-i`) parameter.

This is because yargs 11 started to treat `requiresArg` as an implicit
`nargs: 1`, i.e. to accept exactly one value only.

This patch fixes the issue by dropping `requiresArg` and validating the
number of parameters in the `execute` function instead.