fix: Fixed noReload in configuration files not taking effect
In the run command, `noReload` is defined using yargs's boolean negation
to automatically create the `--no-reload` parameter. This does not
allow setting `noReload` from a configuration file and the valid
option `reload` does not affect the behavior of `run`.

This fix properly sets `noReload` and `noInput` using the values of
yarg options `reload` and `input` after all configuration files are
loaded.