test(functional): Add test for auto-reload after failed load
When auto-reload is enabled, and the developer accidentally breaks the
extension, then web-ext will try to reload the extension, which causes
it to be disabled in the browser. In the case of Chrome, the new
CDP-based method can successfully reload the extension (Chrome 126+).

The fallback for older Chrome does not work. This patch adds test
coverage for that scenario, and updates the fake-chrome-binary to
accurately simulate that scenario.

Tested with real Chrome 75, 131, 138, 139.

CHROME_PATH=/path/to/chrome75/chromium TEST_WEBEXT_USE_REAL_CHROME=1 ./node_modules/.bin/mocha tests/functional/test.cli.run-target-chromium.js